feature/requirements.txt
piexif==1.1.3
pillow==11.2.1
streamlit==1.45.1
pandas==2.2.3
gspread==6.2.1
pydeck==0.9.1
torch==2.7.0
ultralytics==8.3.146
ultralytics-thop==2.0.14
opencv-python==4.11.0.86
